# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Uskup Surabaya API",
    "description": "REST API for Diocese of Surabaya Dashboard Application.\n\n## Authentication\nAll endpoints (except `/api/auth/login`) require JWT authentication. Include the token in cookies or Authorization header.\n\n## Roles\n- **USKUP**: Full access to all features\n- **VIKJEN**: High-level administrative access\n- **SEKRETARIS**: Document and scheduling management\n- **STAF**: Limited access to basic features\n- **VIEWER**: Read-only access",
    "version": "2.9.4",
    "contact": {
      "name": "Keuskupan Surabaya",
      "email": "admin@keuskupan-surabaya.org"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3053",
      "description": "Development server"
    },
    {
      "url": "https://dashboard.keuskupan-surabaya.org",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request password reset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reset email sent (if email exists)" }
        }
      }
    },
...
    "/api/surat/numbering": {
      "get": {
        "tags": ["Surat"],
        "summary": "Get surat numbering",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "action", "in": "query", "schema": { "type": "string", "enum": ["generate", "options", "stats"] } },
          { "name": "jenis", "in": "query", "schema": { "type": "string" } },
          { "name": "unit", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Numbering data" }
        }
      }
    },
...
    "/api/push/subscribe": {
      "get": {
        "tags": ["Push"],
        "summary": "Get VAPID public key",
        "responses": {
          "200": { "description": "VAPID public key for subscription" }
        }
      },
      "post": {
        "tags": ["Push"],
        "summary": "Subscribe to push notifications",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PushSubscription" }
            }
          }
        },
        "responses": {
          "200": { "description": "Subscribed" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PushSubscription": {
        "type": "object",
        "required": ["subscription"],
        "properties": {
          "subscription": {
            "type": "object",
            "properties": {
              "endpoint": { "type": "string" },
              "keys": {
                "type": "object",
                "properties": {
                  "p256dh": { "type": "string" },
                  "auth": { "type": "string" }
                }
              }
            }
          },
          "action": { "type": "string", "enum": ["subscribe", "unsubscribe"] }
        }
      }
    }
  }
}